Rename _logger.py to logger.py
