+++ backend/app/db/session.py
@@ -2,7 +2,7 @@ from collections.abc import AsyncGenerator
 
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
 
-from app.core.config import settings
+from app.config import settings
 
 engine = create_async_engine(
     settings.database_url,
