+++ backend/pyproject.toml
@@ -53,7 +53,6 @@ mypy = "1.18.2"
 torch = "2.7.1"
 ipykernel = "7.0.0"
 
-
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
